back/requirements/common.txt
aioredis==1.3.1
arrow==0.13.1
bleach==3.1.0
boto3==1.9.95
botocore==1.12.151
cacheout==0.11.1
channels ==2.4.0
channels-redis==3.0.1
daphne==2.5.0
diff-match-patch==20120106
dj-database-url==0.5.0
Django==2.2.14
django-filter==2.1.0
django-impersonate==1.4
django-ipware==1.1.1
django-mptt==0.9.1
django-picklefield==2.0
django-redis==4.10.0
django-storages==1.7.1
django-toolbelt==0.0.1
django-webpack-loader==0.6.0
djangorestframework==3.11.0
drf-flex-fields==0.8.6
drf-yasg==1.17.1
elasticsearch==7.0.2
elasticsearch-dsl==7.0.0
lxml==4.3.1
openpyxl==2.6.0
Pillow==7.2.0
python-docx==0.8.10
python-postmark==0.5.1
pytz==2018.9
redis==3.5.3
requests==2.21.0
rq==1.5.0
sentry-sdk==0.15.1
stripe==1.82.1
talon==1.4.4
waitress==1.2.1
whitenoise==3.3.0
xhtml2pdf==0.2.4
